+++ packages/@react-elf/layout/src/flex/Flex.tsx
@@ -1,0 +1,35 @@
+import { propertyMap } from "@react-elf/shared";
+import classNames from "classnames";
+import { useMemo } from "react";
+
+import { FlexProps } from "@react-elf-types/layout";
+
+
+export function Flex (props: FlexProps) {
+    const {
+      style = {},
+      className = "",
+      children,
+      stack,
+      wrap = false,
+      sameWidth = false,
+    } = props;
+
+    const localClass = useMemo(() => {
+      return classNames("elf--flex", className, {
+        stack,
+        wrap,
+        "same-width": sameWidth,
+      });
+    }, [className, stack, wrap, sameWidth]);
+
+    const styleObject = {
+      class: localClass,
+
+      style: {
+        ...propertyMap(style, {}),
+      },
+    };
+
+    return <div {...styleObject}>{children}</div>;
+}