+++ packages/@react-elf/layout/src/grid/Column.tsx
@@ -1,0 +1,16 @@
+import { ColumnProps } from "@react-elf-types/layout";
+
+export function Column({ children, span = 1, style, ...extraProps }: ColumnProps) {
+  return (
+    <div
+      className="elf--column"
+      {...extraProps}
+      style={{
+        ...style,
+        gridColumn: `span ${span}`,
+      } as React.CSSProperties}
+    >
+      {children}
+    </div>
+  );
+}