+++ packages/@react-elf/alert/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @react-elf/button
 
+## 0.0.13
+
+### Patch Changes
+
+- update package.json
+
 ## 0.0.11
 
 ### Patch Changes